# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: 家庭 AI 学习助手 API
  version: 0.6.0
  description: Self-hosted Household account/session and learning contract vertical slices.
servers:
  - url: http://localhost:8000
paths:
...
  /households/{household_id}/captures/{capture_id}/image-analysis-jobs:
    parameters:
      - $ref: '#/components/parameters/HouseholdId'
      - name: capture_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      operationId: startImageAnalysis
      summary: Record a user-confirmed sanitization receipt for one Capture
      description: The service queues a job only when NewAPI is explicitly enabled and the user-confirmed sanitization hash matches the Capture; otherwise it records a blocked job. Only the confirmed derivative may reach the self-hosted Provider.
      security:
        - SessionCookie: []
        - BearerSession: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartImageAnalysisRequest'
      responses:
        '202':
          description: Receipt accepted; status is queued when the self-hosted Provider is enabled, otherwise blocked.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageAnalysisJobReceipt'
        '200':
          description: Idempotent replay of the same receipt.
          headers:
            Idempotency-Replayed:
              schema:
                type: string
                const: 'true'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageAnalysisJobReceipt'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Capture version, upload state, or idempotency conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /households/{household_id}/captures/{capture_id}/image-analysis-jobs/{job_id}:
    parameters:
      - $ref: '#/components/parameters/HouseholdId'
      - name: capture_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getImageAnalysisJob
      summary: Read provider-neutral ImageAnalysis job state
      description: The response contains only lifecycle, receipt hash, and stable error fields; no image, URL, object key, OCR text, or Provider response is returned.
      security:
        - SessionCookie: []
        - BearerSession: []
      responses:
        '200':
          description: ImageAnalysis job state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageAnalysisJobReceipt'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /households/{household_id}/captures/{capture_id}/image-analysis-jobs/{job_id}/extraction:
    parameters:
      - $ref: '#/components/parameters/HouseholdId'
      - name: capture_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getQuestionExtraction
      summary: Read an unverified question extraction for manual review
      description: Returns only schema-validated extraction fields. The result is not a verified question and must not be used as a Tutor fact before human confirmation.
      security:
        - SessionCookie: []
        - BearerSession: []
      responses:
        '200':
          description: Unverified question extraction.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionExtractionRecord'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Extraction has not completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /households/{household_id}/captures/{capture_id}/image-analysis-jobs/{job_id}/extraction/verify:
    parameters:
      - $ref: '#/components/parameters/HouseholdId'
      - name: capture_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      operationId: verifyQuestionExtraction
      summary: Confirm and edit one extraction into a VerifiedQuestion
      description: The unverified extraction remains immutable. Only explicitly submitted fields become a Tutor-safe fact after Household and Capture version checks.
      security:
        - SessionCookie: []
        - BearerSession: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyQuestionRequest'
      responses:
        '201':
          description: Human-confirmed question fact created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifiedQuestion'
        '200':
          description: Idempotent replay of the same confirmation.
          headers:
            Idempotency-Replayed:
              schema:
                type: string
                const: 'true'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifiedQuestion'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Capture version, extraction readiness, or idempotency conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /households/{household_id}/captures/{capture_id}/image-analysis-jobs/{job_id}/verified-question:
    parameters:
      - $ref: '#/components/parameters/HouseholdId'
      - name: capture_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getVerifiedQuestion
      summary: Read a human-confirmed question fact
      security:
        - SessionCookie: []
        - BearerSession: []
      responses:
        '200':
          description: Verified question.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifiedQuestion'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
    HouseholdId:
      name: household_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
...
  securitySchemes:
    BearerSession:
      type: http
      scheme: bearer
      bearerFormat: opaque 256-bit session token
      description: Flutter session token; only a SHA-256 digest is stored by the API.
    SessionCookie:
      type: apiKey
      in: cookie
      name: study_session
      description: HttpOnly SameSite=Lax Web session cookie.
  schemas:
...
    PrivacySanitization:
      type: object
      additionalProperties: false
      required: [schema_version, sanitizer_version, safe_to_upload, requires_confirmation, sensitive_types, region_count, face_detected, qr_detected, barcode_detected, blocked_reasons, sanitized_derivative_sha256]
      properties:
        schema_version:
          type: string
          const: privacy-sanitization.v1
        sanitizer_version:
          type: string
          minLength: 1
          maxLength: 64
        safe_to_upload:
          type: boolean
        requires_confirmation:
          type: boolean
          const: true
        sensitive_types:
          type: array
          uniqueItems: true
          maxItems: 32
          items:
            type: string
            enum: [name, school, class, grade, student_id, exam_id, seat_number, phone, address, signature, face, qr_code, barcode]
        region_count:
          type: integer
          minimum: 0
          maximum: 256
        face_detected:
          type: boolean
        qr_detected:
          type: boolean
        barcode_detected:
          type: boolean
        blocked_reasons:
          type: array
          maxItems: 16
          items:
            type: string
            minLength: 1
            maxLength: 64
        sanitized_derivative_sha256:
          type: string
          pattern: '^[a-f0-9]{64}$'
    StartImageAnalysisRequest:
      type: object
      additionalProperties: false
      required: [expected_capture_version, sanitization, user_confirmed]
      properties:
        expected_capture_version:
          type: integer
          minimum: 1
        sanitization:
          $ref: '#/components/schemas/PrivacySanitization'
        user_confirmed:
          type: boolean
          const: true
    ImageAnalysisJobReceipt:
      type: object
      additionalProperties: false
      required: [id, capture_id, household_id, child_id, status, attempt, sanitization_schema_version, sanitized_derivative_sha256, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        capture_id:
          type: string
          format: uuid
        household_id:
          type: string
          format: uuid
        child_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued, running, succeeded, failed, blocked]
        attempt:
          type: integer
          minimum: 0
        sanitization_schema_version:
          type: string
          const: privacy-sanitization.v1
        sanitized_derivative_sha256:
          type: string
          pattern: '^[a-f0-9]{64}$'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        extraction_id:
          type: string
          format: uuid
          nullable: true
        error_code:
          type: string
          minLength: 1
          maxLength: 64
          nullable: true
    QuestionExtractionRecord:
      type: object
      additionalProperties: false
      required: [id, capture_id, household_id, child_id, extraction, created_at]
      properties:
        id:
          type: string
          format: uuid
        capture_id:
          type: string
          format: uuid
        household_id:
          type: string
          format: uuid
        child_id:
          type: string
          format: uuid
        extraction:
          $ref: '#/components/schemas/QuestionExtraction'
        created_at:
          type: string
          format: date-time
    VerifyQuestionRequest:
      type: object
      additionalProperties: false
      required: [expected_capture_version, question_text, options, formulas, has_diagram, has_handwriting]
      properties:
        expected_capture_version:
          type: integer
          minimum: 1
        question_text:
          type: string
          minLength: 1
          maxLength: 4000
        options:
          type: array
          maxItems: 20
          items:
            type: string
        formulas:
          type: array
          maxItems: 50
          items:
            type: string
        has_diagram:
          type: boolean
        has_handwriting:
          type: boolean
        answer_text:
          type: string
          maxLength: 1000
          nullable: true
    VerifiedQuestion:
      type: object
      additionalProperties: false
      required: [id, capture_id, extraction_id, version, subject, question_text, options, formulas, has_diagram, has_handwriting, verified_by, verified_at]
      properties:
        id:
          type: string
          format: uuid
        capture_id:
          type: string
          format: uuid
        extraction_id:
          type: string
          format: uuid
        version:
          type: integer
          minimum: 1
        subject:
          type: string
          const: math
        question_text:
          type: string
          minLength: 1
          maxLength: 4000
        options:
          type: array
          maxItems: 20
          items:
            type: string
        formulas:
          type: array
          maxItems: 50
          items:
            type: string
        has_diagram:
          type: boolean
        has_handwriting:
          type: boolean
        answer_text:
          type: string
          maxLength: 1000
          nullable: true
        verified_by:
          type: string
          enum: [child, parent]
        verified_at:
          type: string
          format: date-time
...
  responses:
    Unauthorized:
      description: Missing or invalid synthetic principal.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Principal role cannot perform this operation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource is not visible in the caller's household.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...